requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4
google-generativeai==0.3.1
requests==2.31.0
Pillow==10.0.1
numpy==1.24.3
rembg==2.0.50
opencv-python-headless==4.8.1.78
moviepy==1.0.3
imageio==2.31.6
imageio-ffmpeg==0.4.9
decorator==4.4.2
proglog==0.1.10
tqdm==4.66.1
git+https://github.com/openai/whisper.git
onnxruntime==1.16.1